Nomute | Overload the update function signature to serve as an alias for set.
The signature of the update function has been changed so as to allow it use a new value directly instead of using a modification function. This effectively makes the update function an alias of the set function as update uses set internally. This means we can either useupdate or set interchangeably when a new value is supposed to be set for an object key or we can use update by passing in a modification function for more performing more complex updation logic.