Nomute | Add initial version of nestedUpdate | Throw error for invalid object key usage
This commit adds the nestedUpdate function with limited type-safety. TypeScript does not have a straightforward way for statically checking the key types for deeply nested objects. This severely limits the ability to make the nestedUpdate function fully type-s-safe. In order to to mitigate invalid or incorrect updates however, we have added a check for invalid keys in the set function which is lowest level function for updating simple as well as deeply nested objects.